fix: remove o controller e o service de comments de dentro de events
